segregated the preprocess into separate module
